T - Sanctrum USER BBDD(ended)

+++ habitus-api/app/Http/Controllers/UserController.php
@@ -3,9 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\SignupRequest;
+use App\Http\Requests\UpdateUserRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -14,15 +15,26 @@ class UserController extends Controller
      */
     public function index()
     {
-        /**
-         $user = Schema::getColumnListing((new User)->getTable());
 
-         dd($user);
-         *
-         * $user = User::update($request->all());
-         *
-         */
-        return User::get();
+        if (Auth::check() && Auth::user()->rol === 'patient') {
+            // Si el usuario es un paciente, mandamos directiva
+            // 403, sin permisos con un mensaje.
+            return response()->json([
+                'message' => 'No tienes permiso.'
+            ], 403);
+        }
+
+        if (Auth::check() && Auth::user()->rol === 'psychologist') {
+            // los psicologos solo podrán ver a los pacientes
+            $users = User::where('rol', 'patient')->get();
+        } else {
+            // Los recepcionistas y administradores lo podrán ver todo
+            $users = User::all();
+        }
+
+        // devolvemos todos los usuarios y controlamos las peticiones http
+        // 200 -> leer
+        return response()->json($users, 200);
     }
 
     /**
@@ -31,6 +43,27 @@ class UserController extends Controller
     public function store(SignupRequest $request)
     {
 
+        if (Auth::check() && Auth::user()->rol === 'patient') {
+            // validamos que si el usuario es paciente, no pueda
+            // crear usuarios.
+            return response()->json([
+                'message' => 'No tienes permiso.',
+            ], 403);
+        }
+        // este $request->validated() valida directaemnte
+        // todos los atributos del request.
+        $data = $request->validated();
+
+        // creamos el usuario y lo guardamos en la bbdd,
+        // esto es posible gracias a no hacer todos los request
+        $user = User::create($data);
+
+        // creamos el usuario y envamos un mensaje de confirmación
+        // 201 -> crear
+        return response()->json([
+            'message' => 'Usuario creado correctamente.',
+            'data'    => $user,
+        ], 201);
     }
 
     /**
@@ -38,19 +71,46 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
-        return $user;
+        if( Auth::check() && Auth::user()->rol === 'patient' ){
+            // Si el usuario es paciente, no tiene acceso al show.
+            return response()->json([
+                'message' => 'No tienes permiso.'
+            ], 403);
+        }
+
+        if(Auth::check() && Auth::user()->rol === 'psychologist' && $user->rol !== 'patient'){
+            return response()->json([
+                'message' => 'No tienes permiso.'
+            ], 403);
+        }
+
+        return response()->json($user, 200);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, User $user)
+    public function update(UpdateUserRequest $request, User $user)
     {
-        //return $request->name;
+        if (Auth::check() && Auth::user()->rol === 'patient') {
+            // validamos que si el usuario es paciente, no pueda
+            // editar usuarios.
+            return response()->json([
+                'message' => 'No tienes permiso.',
+            ], 403);
+        }
 
-        $user->update($request->all());
+        // aquí válidamos de nuevo los datos
+        $data = $request->validated();
 
-        return $user;
+        // y actualizamos directamente con ->update
+        $user->update($data);
+
+        // enviamos un mensaje de confirmación con códido HTTP 200 leer
+        return response()->json([
+            'message' => 'Usuario actualizado correctamente.',
+            'data'    => $user,
+        ], 200);
     }
 
     /**
@@ -58,6 +118,19 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        return $user->delete();
+        if (Auth::check() && Auth::user()->rol === 'patient') {
+            // validamos que si el usuario es paciente, no pueda
+            // borrar usuarios.
+            return response()->json([
+                'message' => 'No tienes permiso.',
+            ], 403);
+        }
+        // boramos el usuario
+        $user->delete();
+
+        // enviamos un mensaje de confirmación con códido HTTP 200 leer
+        return response()->json([
+            'message' => 'Usuario eliminado correctamente.',
+        ], 200);
     }
 }
